pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "unstructured>=0.18.13",
]
